[GStreamer][PipeWire] getDisplayMedia() broken https://bugs.webkit.org/show_bug.cgi?id=283298
Reviewed by Xabier Rodriguez-Calvar.

Move the DMABuf caps builder utility function to GStreamerCommon and use it from the VideoCapturer,
fixing caps negotiation issues with PipeWire.

* Source/WebCore/platform/graphics/gstreamer/GLVideoSinkGStreamer.cpp:
(initializeDMABufAvailability):
(drmFourccToGstVideoFormat): Deleted.
(buildDMABufCaps): Deleted.
* Source/WebCore/platform/graphics/gstreamer/GStreamerCommon.cpp:
(WebCore::drmFourccToGstVideoFormat):
(WebCore::buildDMABufCaps):
* Source/WebCore/platform/graphics/gstreamer/GStreamerCommon.h:
* Source/WebCore/platform/mediastream/gstreamer/GStreamerVideoCapturer.cpp:
(WebCore::GStreamerVideoCapturer::createConverter):

Canonical link: https://commits.webkit.org/286790@main
